# src/main/resources/static/openapi.yaml
openapi: 3.0.3 # version of the specification
info:
  version: '1'
  title: Emerion Clientes API - Meus Pedidos

servers:
  - url: http://localhost:8080

tags:
  - name: CustomerData
    description: Operations over customer basic data
  - name: CustomerCredit
    description: Operations over customer credit data
  - name: CustomerAddress
    description: Operations over customer address data
  - name: CustomerDashboard
    description: Operations over customer dashboard
  - name: CustomerOrderHeader
    description: Operations over customer order header
paths:
  /customer-data:
    post:
      tags:
        - CustomerData
      description: post a customer info data into the database
      operationId: saveCustomerData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerDataDTO"
      responses:
        201:
          description: Customer info successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDataDTO"
    get:
      tags:
        - CustomerData
      description: get all the customer data from the database
      operationId: getAllCustomerData
      responses:
        200:
          description: Customer info successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDataDTOList"

  /customer-data/{customerId}:
    get:
      tags:
        - CustomerData
      description: get customer data from the database
      operationId: getCustomerData
      parameters:
        - name: customerId
          in: path
          description: customerData id
          required: true
          schema:
            $ref: "#/components/schemas/Identifier"
      responses:
        200:
          description: Customer info successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDataDTO"

  ### Customer Credit that has been generated by some returns, for instance
  /customer-credit:
    post:
      tags:
        - CustomerCredit
      description: post a customer credit data into the database
      operationId: saveCustomerCredit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerCreditDTO"
      responses:
        201:
          description: Customer credit successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerCreditDTO"
    get:
      tags:
        - CustomerCredit
      description: get all the customer credit from the database
      operationId: getAllCustomerCredit
      responses:
        200:
          description: Customer credit successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerCreditDTOList"

  /customer-credit/{creditId}:
      get:
        tags:
          - CustomerCredit
        description: get customer credit from the database
        operationId: getCustomerCreditById
        parameters:
          - name: creditId
            in: path
            description: customerCredit id
            required: true
            schema:
              $ref: "#/components/schemas/Identifier"
        responses:
          200:
            description: Customer credit successfully retrieved
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/CustomerCreditDTO"

  ### Customer Address that has been generated by some returns, for instance
  /customer-address:
    post:
      tags:
        - CustomerAddress
      description: post a customer address data into the database
      operationId: saveCustomerAddress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerAddressDTO"
      responses:
        201:
          description: Customer address successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAddressResponseListDTO"
    get:
      tags:
        - CustomerAddress
      description: get all the customer address from the database
      operationId: getAllCustomerAddress
      responses:
        200:
          description: Customer address successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAddressResponseListDTO"

  /customer-address/{customerAddressId}:
    get:
      tags:
        - CustomerAddress
      description: get the address for the specified customer
      operationId: getCustomerAddressById
      parameters:
        - name: customerAddressId
          in: path
          description: customerAddress id
          required: true
          schema:
            $ref: "#/components/schemas/Identifier"
      responses:
        200:
          description: Customer address successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAddressResponseDTO"

  ### Customer Dashboard info
  /customer-dashboard/{customerId}:
    get:
      tags:
        - CustomerDashboard
      description: get the dashboard for the specified customer
      operationId: getCustomerDashboard
      parameters:
        - name: customerId
          in: path
          description: customerDashboard id
          required: true
          schema:
            $ref: "#/components/schemas/Identifier"
      responses:
        200:
          description: Customer dashboard successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDashboardInfoDTO"

  ### Customer Order that has been generated by some returns, for instance
  /customer-order/header:
    post:
      tags:
        - CustomerOrder
      description: post a customer order data into the database
      operationId: saveCustomerOrderHeader
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerOrderHeaderDTO"
      responses:
        201:
          description: Customer order successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrderHeaderDTO"
    get:
      tags:
        - CustomerOrder
      description: get all the customer order from the database
      operationId: getAllCustomerOrderHeader
      responses:
        200:
          description: Customer order successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrderHeaderDTOList"

  /customer-order/header/{orderId}:
    get:
      tags:
        - CustomerOrder
      description: get the order for the specified customer
      operationId: getCustomerOrderHeaderByOrderId
      parameters:
        - name: orderId
          in: path
          description: order id
          required: true
          schema:
            $ref: "#/components/schemas/Identifier"
      responses:
        200:
          description: Customer order successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrderHeaderDTO"

  /customer-order/header/customer/{customerId}:
    get:
      tags:
        - CustomerOrder
      description: get the order for the specified customer
      operationId: getCustomerOrderHeaderByCustomer
      parameters:
        - name: customerId
          in: path
          description: customer id
          required: true
          schema:
            $ref: "#/components/schemas/Identifier"
      responses:
        200:
          description: Customer order successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrderHeaderDTOList"

  /customer-order/detail:
    post:
      tags:
        - CustomerOrder
      description: post a customer order detail data into the database
      operationId: saveCustomerOrderDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerOrderDetailSummaryDTO"
      responses:
        201:
          description: Customer order detail successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrderDetailSummaryDTO"
    get:
      tags:
        - CustomerOrder
      description: get all the customer order detail from the database
      operationId: getAllCustomerOrderDetail
      responses:
        200:
          description: Customer order detail successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrderDetailSummaryDTOList"

  /customer-order/detail/customer/{customerId}:
    get:
      tags:
        - CustomerOrder
      description: get the order for the specified customer
      operationId: getCustomerOrderDetailByCustomer
      parameters:
        - name: customerId
          in: path
          description: customer id
          required: true
          schema:
            $ref: "#/components/schemas/Identifier"
      responses:
        200:
          description: Customer Order detail successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrderDetailSummaryDTOList"

  /customer-order/detail/order/{orderId}:
    get:
      tags:
        - CustomerOrder
      description: get the order for the specified customer
      operationId: getCustomerOrderDetailByOrderId
      parameters:
        - name: orderId
          in: path
          description: order id
          required: true
          schema:
            $ref: "#/components/schemas/Identifier"
      responses:
        200:
          description: Customer Order detail successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrderDetailSummaryDTOList"

components:
  schemas:
  ### Complex types
    Double:
      type: number
      format: double
    Instant:
      type: string
      format: date-time

    Identifier:
      type: string
      format: uuid

  ### Objects
    CustomerDataDTOList:
      type: array
      items:
        $ref: "#/components/schemas/CustomerDataDTO"
    CustomerDataDTO:
        type: object
        properties:
          id:
            $ref: "#/components/schemas/Identifier"
          cnpjEmpresa:
              type: string
          codcli:
            type: string
          nomcli:
            type: string
          apecli:
            type: string
          cgccli:
            type: string
          inscli:
            type: string
          nomregtrib:
            type: string
          clidev:
            type: string
          cifcli:
            type: string
          uffcli:
            type: string
          codven:
            type: string
          cnae:
            type: string
          indicEstrangeiro:
            type: string
          tipoIndicacaoIe:
            type: string
          inscMunicipal:
            type: string
          obscli:
            type: string
          indicIe:
            type: integer
    CustomerCreditDTOList:
      type: array
      items:
        $ref: "#/components/schemas/CustomerCreditDTO"
    CustomerCreditDTO:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Identifier"
        cnpjEmpresa:
          type: string
        codcli:
          type: string
        seqcde:
          type: string
        dtecde:
          $ref: "#/components/schemas/Instant"
        tipo:
          type: string
        valor:
          $ref: "#/components/schemas/Double"
        usacde:
          $ref: "#/components/schemas/Double"
        valcde:
          $ref: "#/components/schemas/Double"
        sldcde:
          $ref: "#/components/schemas/Double"
        dteped:
          type: string
        obscde:
          type: string
        sitcde:
          type: string
    CustomerAddressDTO:
      type: object
      properties:
        cnpjEmpresa:
          type: string
        codcli:
          type: string
        addressList:
          type: array
          items:
            $ref: "#/components/schemas/CustomerAddressDetailDTO"
    CustomerAddressDetailDTO:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Identifier"
        tipo:
          type: string
        cefcli:
          type: string
        tefcli:
          type: string
        enfcli:
          type: string
        nrfcli:
          type: string
        rffcli:
          type: string
        bafcli:
          type: string
        cifcli:
          type: string
        uffcli:
          type: string
        pt1cli:
          type: string
        fo1cli:
          type: string
        cofcli:
          type: string
        pc1cli:
          type: string
        fc1cli:
          type: string
    CustomerAddressResponseListDTO:
      type: array
      items:
        $ref: "#/components/schemas/CustomerAddressResponseDTO"
    CustomerAddressResponseDTO:
      type: object
      properties:
        cnpjEmpresa:
          type: string
        codcli:
          type: string
        id:
          $ref: "#/components/schemas/Identifier"
        tipo:
          type: string
        cefcli:
          type: string
        tefcli:
          type: string
        enfcli:
          type: string
        nrfcli:
          type: string
        rffcli:
          type: string
        bafcli:
          type: string
        cifcli:
          type: string
        uffcli:
          type: string
        pt1cli:
          type: string
        fo1cli:
          type: string
        cofcli:
          type: string
        pc1cli:
          type: string
        fc1cli:
          type: string
    CustomerDashboardInfoDTO:
      type: object
      properties:
        customerCreditList:
          $ref: "#/components/schemas/CustomerCreditDTOList"
        customerData:
          $ref: "#/components/schemas/CustomerDataDTO"
        customerAddress:
          $ref: "#/components/schemas/CustomerAddressResponseListDTO"
    CustomerOrderHeaderDTOList:
      type: array
      items:
        $ref: "#/components/schemas/CustomerOrderHeaderDTO"
    CustomerOrderHeaderDTO:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Identifier"
        cnpjEmpresa:
          type: string
        codcli:
          type: string
        numres:
          type: string
        dteres:
          $ref: "#/components/schemas/Instant"
        sitres:
          type: string
        totger:
          $ref: "#/components/schemas/Double"
        totipi:
          $ref: "#/components/schemas/Double"
        totsub:
          $ref: "#/components/schemas/Double"
        totdescinc:
          $ref: "#/components/schemas/Double"
        totres:
          $ref: "#/components/schemas/Double"
    CustomerOrderDetailSummaryDTOList:
      type: array
      items:
        $ref: "#/components/schemas/CustomerOrderDetailSummaryDTO"
    CustomerOrderDetailSummaryDTO:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Identifier"
        cnpjEmpresa:
          type: string
        codcli:
          type: string
        numres:
          type: string
        desre2:
          type: string
        vlqre2:
          $ref: "#/components/schemas/Double"
        qtpre2:
          $ref: "#/components/schemas/Double"
        totre2:
          $ref: "#/components/schemas/Double"
        totger:
          $ref: "#/components/schemas/Double"